requirements.txt
colorzero==2.0
gpiozero==1.6.1
pkg-resources==0.0.0
RPi.GPIO==0.7.0
